+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The following changes were made after the previous version, and will be availabl
 - rtmidi.GetMessageType
 - rtmidi.GetMessageChannel
 - rtmidi.GetControlName
+- rtmidi.GetControlBitSignificance
 
 #### Enumerations